+++ packages/liveblocks-core/src/room.ts
@@ -662,6 +662,13 @@ export type Room<
     readonly storageStatus: Observable<StorageStatus>;
     readonly ydoc: Observable<YDocUpdateServerMsg | UpdateYDocClientMsg>;
     readonly comments: Observable<CommentsEventServerMsg>;
+
+    /**
+     * Called right before the room is destroyed. The event cannot be used to
+     * prevent the room from being destroyed, only to be informed that this is
+     * imminent.
+     */
+    readonly roomWillDestroy: Observable<void>;
   };
 
   /**
@@ -1590,6 +1597,7 @@ export function createRoom<
     ydoc: makeEventSource<YDocUpdateServerMsg | UpdateYDocClientMsg>(),
 
     comments: makeEventSource<CommentsEventServerMsg>(),
+    roomWillDestroy: makeEventSource<void>(),
   };
 
   const roomId = config.roomId;
@@ -2829,6 +2837,7 @@ export function createRoom<
     ydoc: eventHub.ydoc.observable,
 
     comments: eventHub.comments.observable,
+    roomWillDestroy: eventHub.roomWillDestroy.observable,
   };
 
   async function getThreadsSince(options: GetThreadsSinceOptions) {
@@ -3089,11 +3098,21 @@ export function createRoom<
       reconnect: () => managedSocket.reconnect(),
       disconnect: () => managedSocket.disconnect(),
       destroy: () => {
-        syncSourceForStorage.destroy();
+        // remove the roomWillDestroy event from the event hub
+        const { roomWillDestroy, ...eventsExceptDestroy } = eventHub;
+        // Unregister all registered callbacks
+        for (const source of Object.values(eventsExceptDestroy)) {
+          source[Symbol.dispose]();
+        }
+        eventHub.roomWillDestroy.notify();
         context.yjsProvider?.off("status", yjsStatusDidChange);
+        syncSourceForStorage.destroy();
         syncSourceForYjs.destroy();
         uninstallBgTabSpy();
         managedSocket.destroy();
+
+        // cleanup will destroy listener
+        roomWillDestroy[Symbol.dispose]();
       },
 
       // Presence