+++ packages/liveblocks-react-tiptap/src/types.ts
@@ -18,6 +18,8 @@ export const LIVEBLOCKS_MENTION_PASTE_KEY = new PluginKey(
 export const LIVEBLOCKS_MENTION_NOTIFIER_KEY = new PluginKey(
   "lb-plugin-mention-notify"
 );
+
+export const LIVEBLOCKS_MENTION_EXTENSION = "liveblocksMentionExt";
 export const LIVEBLOCKS_MENTION_TYPE = "liveblocksMention";
 
 export const THREADS_ACTIVE_SELECTION_PLUGIN = new PluginKey(
@@ -91,7 +93,7 @@ export type LiveblocksExtensionOptions = {
 export type LiveblocksExtensionStorage = {
   unsubs: (() => void)[];
   doc: Doc;
-  provider: LiveblocksYjsProvider<any, any, any, any, any>;
+  provider: LiveblocksYjsProvider;
   permanentUserData: PermanentUserData;
 };